added environment file for prod and renamed environment.ts + added arg with api key in the dokerfile + updated CI with action to triger deployment
