+++ src/environments/environment.dev.ts
@@ -2,4 +2,5 @@ export const environment = {
   production: false,
   mapboxToken: '${{ secrets.MAPBOX_ACCESS_TOKEN }}',
   stripeToken: '${{ secrets.STRIPE_TOKEN }}',
+  apiUrl: 'http://localhost:8081/api'
 };
