+++ src/environments/environment.prod.ts
@@ -1,0 +1,6 @@
+export const environment = {
+  production: true,
+  mapboxToken: 'MAPBOX_TOKEN_PLACEHOLDER',
+  stripeToken: 'STRIPE_TOKEN_PLACEHOLDER',
+  apiUrl: 'http://51.44.22.28:9003/api'
+};
